+++ src/main/java/com/baichen/rpc/compressor/Compressor.java
@@ -1,21 +1,8 @@
 package com.baichen.rpc.compressor;
 
-import lombok.Getter;
+import com.baichen.rpc.spi.Extension;
 
-public interface Compressor {
+public interface Compressor extends Extension {
     byte[] compress(byte[] data);
     byte[] decompress(byte[] data);
-
-    @Getter
-    enum CompressorType {
-        NONE(1),
-        GZIP(2),
-        ;
-
-        private final byte code;
-
-        CompressorType(int code) {
-            this.code = (byte) code;
-        }
-    }
 }