+++ src/main/java/com/baichen/rpc/compressor/NoneCompressor.java
@@ -1,5 +1,7 @@
 package com.baichen.rpc.compressor;
 
+import java.util.Locale;
+
 public class NoneCompressor implements Compressor {
     @Override
     public byte[] compress(byte[] data) {
@@ -9,5 +11,15 @@ public class NoneCompressor implements Compressor {
     @Override
     public byte[] decompress(byte[] data) {
         return data;
+    }
+
+    @Override
+    public String getName() {
+        return "none";
+    }
+
+    @Override
+    public int getCode() {
+        return 0;
     }
 }