+++ src/main/java/com/baichen/rpc/serializer/JsonSerializer.java
@@ -4,6 +4,7 @@ import com.alibaba.fastjson2.JSONObject;
 import com.alibaba.fastjson2.JSONReader;
 
 import java.nio.charset.StandardCharsets;
+import java.util.Locale;
 
 public class JsonSerializer implements Serializer {
     @Override
@@ -14,5 +15,15 @@ public class JsonSerializer implements Serializer {
     @Override
     public <T> T deserialize(byte[] bytes, Class<T> clazz) {
         return JSONObject.parseObject(new String(bytes, StandardCharsets.UTF_8), clazz, JSONReader.Feature.SupportClassForName);
+    }
+
+    @Override
+    public String getName() {
+        return "json";
+    }
+
+    @Override
+    public int getCode() {
+        return 0;
     }
 }