+++ src/main/java/com/baichen/rpc/serializer/Serializer.java
@@ -1,23 +1,11 @@
 package com.baichen.rpc.serializer;
 
+import com.baichen.rpc.spi.Extension;
 import lombok.Getter;
 
-public interface Serializer {
+public interface Serializer extends Extension {
 
     byte[] serialize(Object obj);
 
     <T> T deserialize(byte[] bytes, Class<T> clazz);
-
-    @Getter
-    enum SerializerType {
-        JSON(1),
-        HESSIAN(2),
-        ;
-
-        private final byte code;
-
-        SerializerType(int code) {
-            this.code = (byte) code;
-        }
-    }
 }