+++ src/main/java/com/baichen/rpc/spi/Extension.java
@@ -1,0 +1,10 @@
+package com.baichen.rpc.spi;
+
+public interface Extension {
+
+    String getName();
+
+    default int getCode() {
+        return -1;
+    }
+}